refactor: 패키지 명 exception 으로 변경
