--- src/main/java/com/mju/lighthouseai/domain/restaurant/dto/controller/RestaurantReadAllControllerResquestDto.java
@@ -1,14 +1,0 @@
-package com.mju.lighthouseai.domain.restaurant.dto.controller;
-
-public record RestaurantReadAllControllerResquestDto (
-        String title,
-        String location,
-        int price,
-        String menu,
-        String opentime,
-        String closetime,
-        String constituency_name,
-        String nickname,
-        String region_name
-) {
-}
